Yall I tried
Adding info the the cards is quite simple. First, get items by PantryProxy().getPantryItems() to show items which are currently located in the pantry. This is how all functions are called,  with the proxy. This package called expansion_tile_card seemed the prettiest and most flexible option to make the expandable item cards. However, I struggle updating the page state and with height constraints of containers.